remocao_fundo_imagens/requirements.txt
gradio
transformers
Pillow
torchvision
scikit-image
